Refactor app service wizard to have separate prompt/execute steps
