+++ appservice/src/createAppService/SiteNameStep.ts
@@ -6,16 +6,17 @@
 // tslint:disable-next-line:no-require-imports
 import WebSiteManagementClient = require('azure-arm-website');
 import { ResourceNameAvailability } from 'azure-arm-website/lib/models';
-import { AzureNameStep, IAzureNamingRules, IAzureUserInput, resourceGroupNamingRules, ResourceGroupStep, storageAccountNamingRules, StorageAccountStep } from 'vscode-azureextensionui';
+import { AzureNameStep, IAzureNamingRules, IAzureUserInput, ResourceGroupListStep, resourceGroupNamingRules, StorageAccountListStep, storageAccountNamingRules } from 'vscode-azureextensionui';
 import { AppKind, getAppKindDisplayName } from './AppKind';
-import { appServicePlanNamingRules, AppServicePlanStep } from './AppServicePlanStep';
+import { AppServicePlanListStep } from './AppServicePlanListStep';
+import { appServicePlanNamingRules } from './AppServicePlanNameStep';
 import { IAppServiceWizardContext } from './IAppServiceWizardContext';
 
 export class SiteNameStep extends AzureNameStep<IAppServiceWizardContext> {
     public async prompt(wizardContext: IAppServiceWizardContext, ui: IAzureUserInput): Promise<IAppServiceWizardContext> {
         const client: WebSiteManagementClient = new WebSiteManagementClient(wizardContext.credentials, wizardContext.subscriptionId);
-        wizardContext.siteName = (await ui.showInputBox({
-            prompt: `Enter a globally unique name for the new ${getAppKindDisplayName(wizardContext.appKind)}.`,
+        wizardContext.newSiteName = (await ui.showInputBox({
+            prompt: `Enter a globally unique name for the new ${getAppKindDisplayName(wizardContext.newSiteKind)}.`,
             validateInput: async (value: string): Promise<string | undefined> => {
                 value = value ? value.trim() : '';
                 const nameAvailability: ResourceNameAvailability = await client.checkNameAvailability(value, 'site');
@@ -28,26 +29,22 @@ export class SiteNameStep extends AzureNameStep<IAppServiceWizardContext> {
         })).trim();
 
         const namingRules: IAzureNamingRules[] = [resourceGroupNamingRules];
-        if (wizardContext.appKind === AppKind.functionapp) {
+        if (wizardContext.newSiteKind === AppKind.functionapp) {
             namingRules.push(storageAccountNamingRules);
         } else {
             namingRules.push(appServicePlanNamingRules);
         }
-        wizardContext.relatedNameTask = this.generateRelatedName(wizardContext, wizardContext.siteName, namingRules);
+        wizardContext.relatedNameTask = this.generateRelatedName(wizardContext, wizardContext.newSiteName, namingRules);
 
-        return wizardContext;
-    }
-
-    public async execute(wizardContext: IAppServiceWizardContext): Promise<IAppServiceWizardContext> {
         return wizardContext;
     }
 
     protected async isRelatedNameAvailable(wizardContext: IAppServiceWizardContext, name: string): Promise<boolean> {
-        const tasks: Promise<boolean>[] = [ResourceGroupStep.isNameAvailable(wizardContext, name)];
-        if (wizardContext.appKind === AppKind.functionapp) {
-            tasks.push(StorageAccountStep.isNameAvailable(wizardContext, name));
+        const tasks: Promise<boolean>[] = [ResourceGroupListStep.isNameAvailable(wizardContext, name)];
+        if (wizardContext.newSiteKind === AppKind.functionapp) {
+            tasks.push(StorageAccountListStep.isNameAvailable(wizardContext, name));
         } else {
-            tasks.push(AppServicePlanStep.isNameAvailable(wizardContext, name, name));
+            tasks.push(AppServicePlanListStep.isNameAvailable(wizardContext, name, name));
         }
 
         return (await Promise.all(tasks)).every((v: boolean) => v);
