+++ appservice/src/createAppService/SiteRuntimeStep.ts
@@ -1,0 +1,118 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { AzureWizardPromptStep, IAzureQuickPickItem } from 'vscode-azureextensionui';
+import { IAzureUserInput } from 'vscode-azureextensionui';
+import { WebsiteOS } from './AppKind';
+import { IAppServiceWizardContext } from './IAppServiceWizardContext';
+
+interface ILinuxRuntimeStack {
+    name: string;
+    displayName: string;
+}
+
+export class SiteRuntimeStep extends AzureWizardPromptStep<IAppServiceWizardContext> {
+    public async prompt(wizardContext: IAppServiceWizardContext, ui: IAzureUserInput): Promise<IAppServiceWizardContext> {
+        if (!wizardContext.newSiteRuntime && wizardContext.newSiteOS === WebsiteOS.linux) {
+            const runtimeItems: IAzureQuickPickItem<ILinuxRuntimeStack>[] = this.getLinuxRuntimeStack().map((rt: ILinuxRuntimeStack) => {
+                return {
+                    id: rt.name,
+                    label: rt.displayName,
+                    description: '',
+                    data: rt
+                };
+            });
+
+            wizardContext.newSiteRuntime = (await ui.showQuickPick(runtimeItems, { placeHolder: 'Select a runtime for your new Linux app.' })).data.name;
+        }
+
+        return wizardContext;
+    }
+
+    private getLinuxRuntimeStack(): ILinuxRuntimeStack[] {
+        return [
+            {
+                name: 'node|4.4',
+                displayName: 'Node.js 4.4'
+            },
+            {
+                name: 'node|4.5',
+                displayName: 'Node.js 4.5'
+            },
+            {
+                name: 'node|6.2',
+                displayName: 'Node.js 6.2'
+            },
+            {
+                name: 'node|6.6',
+                displayName: 'Node.js 6.6'
+            },
+            {
+                name: 'node|6.9',
+                displayName: 'Node.js 6.9'
+            },
+            {
+                name: 'node|6.10',
+                displayName: 'Node.js 6.10'
+            },
+            {
+                name: 'node|6.11',
+                displayName: 'Node.js 6.11'
+            },
+            {
+                name: 'node|8.0',
+                displayName: 'Node.js 8.0'
+            },
+            {
+                name: 'node|8.1',
+                displayName: 'Node.js 8.1'
+            },
+            {
+                name: 'node|8.2',
+                displayName: 'Node.js 8.2'
+            },
+            {
+                name: 'node|8.8',
+                displayName: 'Node.js 8.8'
+            },
+            {
+                name: 'node|8.9',
+                displayName: 'Node.js 8.9 (LTS - Recommended for new apps)'
+            },
+            {
+                name: 'node|9.4',
+                displayName: 'Node.js 9.4'
+            },
+            {
+                name: 'php|5.6',
+                displayName: 'PHP 5.6'
+            },
+            {
+                name: 'php|7.0',
+                displayName: 'PHP 7.0'
+            },
+            {
+                name: 'php|7.2',
+                displayName: 'PHP 7.2'
+            },
+            {
+                name: 'dotnetcore|1.0',
+                displayName: '.NET Core 1.0'
+            },
+            {
+                name: 'dotnetcore|1.1',
+                displayName: '.NET Core 1.1'
+            },
+            {
+                name: 'dotnetcore|2.0',
+                displayName: '.NET Core 2.0'
+            },
+            {
+                name: 'ruby|2.3',
+                displayName: 'Ruby 2.3'
+            }
+        ];
+    }
+}
